// src/types.ts
export type CommunityCategory =
  | 'Ask Clear Clause'
  | 'Privacy'
  | 'Refund & Money'
  | 'Hidden Clauses'
  | 'Security'
  | 'App/Website Reviews'
  | 'Risk Alerts'
  | 'Learn'
  | 'Suggestions'
  | 'General Discussion';